mcp: pat_min_role gate, strict chart passthrough, polish (cobra, split, role consts), docs.
