+++ CHANGELOG.md
@@ -3,8 +3,8 @@
 ## Unreleased
 
 ### Features
-- **Personal access tokens (PAT):** a logged-in user can mint per-user API tokens (`POST`/`GET`/`DELETE /api/auth/tokens`, plus a *Personal access tokens* dialog in the user menu) used as the `X-API-Key` header — bearer credentials for non-browser clients (scripts, the upcoming MCP server) that preserve the user's role (Casbin RBAC). Tokens are stored hashed in the snapshot DB (`api_tokens` table, created by `dasha migrate`), with least-privilege (requested role ≤ caller's, default `viewer`), anti-chaining (cannot mint from a PAT), a one-time secret reveal, immediate revoke, and optional expiry. Token resolution is cached with a short TTL to keep the auth hot path off the database, and `last_used` is updated at most once per token per interval.
-- **MCP connector (`dasha-mcp`):** a separate, read-only [MCP](https://modelcontextprotocol.io) server over the Dasha API for AI assistants (Claude Desktop / Code, opencode, Cursor, …). 22 read-only tools (clusters, fleet health ranking, instance info, health score/trend/per-database & recommendations, top/running/blocked queries, query report & snapshot compare, indexes, top tables, table drill-down, replication, settings analysis, wait events, connections, vacuum-wraparound danger, Yandex Cloud log search — rate-limit-aware: dedup-by-default, local argument validation so a bad request never burns a rate-limit slot, and 429/501/504 mapped to actionable guidance for the model) and 5 prompts (`diagnose_cluster`, `explain_health_score`, `find_index_opportunities`, `investigate_slow_queries`, `fleet_overview`). Every tool carries the MCP read-only + closed-world annotations so clients can auto-approve it; the server sends usage instructions that prime the model, and HTTP mode shares one schema cache across per-token servers. Runs over **stdio** (single identity) or **HTTP/SSE** with per-user token passthrough (no shared server token — Casbin RBAC preserved). Hardened: per-tool result size cap (oversized results refused, never truncated to invalid JSON), optional `limit` args on `describe_table` (partitions) and `connections` (activity sample), hashed + bounded per-token server cache, tokens never logged; rate limiting applies upstream via Dasha's per-identity limiter. Ships a bilingual (en/ru, `--lang` / `DASHA_MCP_LANG`) embedded knowledge base as 3 MCP resources (`dasha://kb/health-rules` — every scoring rule with LOW/MED/HIGH thresholds and first actions, kept in lockstep with the rules engine by a sync test; `dasha://kb/wait-events` — wait event glossary; `dasha://kb/workflow` — complaint-to-tool-chain playbooks and API care rules), and the 5 prompts are linear playbooks (numbered steps, one tool per step, an interpretation criterion on each) — so LLMs without deep PostgreSQL expertise can drive the tools competently, while strong models are not slowed down. Ops parity with the backend: every MCP call is logged via zap to stderr (method, tool, duration — never arguments or tokens), HTTP mode drains in-flight requests on SIGTERM/SIGINT, and the advertised server version is stamped from the release build (same internal/version ldflags scheme as the backend). Typed Dasha API client generated from the OpenAPI spec; multi-arch image `deploy/images/Dockerfile.mcp`, published in the release workflow; optional gated Helm Deployment + Service (`mcp.enabled`).
+- **Personal access tokens (PAT):** a logged-in user can mint per-user API tokens (`POST`/`GET`/`DELETE /api/auth/tokens`, plus a *Personal access tokens* dialog in the user menu) used as the `X-API-Key` header — bearer credentials for non-browser clients (scripts, the upcoming MCP server) that preserve the user's role (Casbin RBAC). Tokens are stored hashed in the snapshot DB (`api_tokens` table, created by `dasha migrate`), with least-privilege (requested role ≤ caller's, default `viewer`), anti-chaining (cannot mint from a PAT), a one-time secret reveal, immediate revoke, and optional expiry. Management is additionally gated by `auth.pat_min_role` (`admin` by default while the feature matures; set `viewer` to open it to every signed-in user). Token resolution is cached with a short TTL to keep the auth hot path off the database, and `last_used` is updated at most once per token per interval.
+- **MCP connector (`dasha-mcp`):** a separate, read-only [MCP](https://modelcontextprotocol.io) server over the Dasha API for AI assistants (Claude Desktop / Code, opencode, Cursor, …). 22 read-only tools (clusters, fleet health ranking, instance info, health score/trend/per-database & recommendations, top/running/blocked queries, query report & snapshot compare, indexes, top tables, table drill-down, replication, settings analysis, wait events, connections, vacuum-wraparound danger, Yandex Cloud log search — rate-limit-aware: dedup-by-default, local argument validation so a bad request never burns a rate-limit slot, and 429/501/504 mapped to actionable guidance for the model) and 5 prompts (`diagnose_cluster`, `explain_health_score`, `find_index_opportunities`, `investigate_slow_queries`, `fleet_overview`). Every tool carries the MCP read-only + closed-world annotations so clients can auto-approve it; the server sends usage instructions that prime the model, and HTTP mode shares one schema cache across per-token servers. Runs over **stdio** (single identity) or **HTTP/SSE** with per-user token passthrough (no shared server token — Casbin RBAC preserved). Hardened: per-tool result size cap (oversized results refused, never truncated to invalid JSON), optional `limit` args on `describe_table` (partitions) and `connections` (activity sample), hashed + bounded per-token server cache, tokens never logged; rate limiting applies upstream via Dasha's per-identity limiter. Ships a bilingual (en/ru, `--lang` / `DASHA_MCP_LANG`) embedded knowledge base as 3 MCP resources (`dasha://kb/health-rules` — every scoring rule with LOW/MED/HIGH thresholds and first actions, kept in lockstep with the rules engine by a sync test; `dasha://kb/wait-events` — wait event glossary; `dasha://kb/workflow` — complaint-to-tool-chain playbooks and API care rules), and the 5 prompts are linear playbooks (numbered steps, one tool per step, an interpretation criterion on each) — so LLMs without deep PostgreSQL expertise can drive the tools competently, while strong models are not slowed down. Ops parity with the backend: every MCP call is logged via zap to stderr (method, tool, duration — never arguments or tokens), HTTP mode drains in-flight requests on SIGTERM/SIGINT, and the advertised server version is stamped from the release build (same internal/version ldflags scheme as the backend). Typed Dasha API client generated from the OpenAPI spec; multi-arch image `deploy/images/Dockerfile.mcp`, published in the release workflow; optional gated Helm Deployment + Service (`mcp.enabled`); the chart deliberately offers no shared fallback token — HTTP mode is strict per-user passthrough.
 
 ## v1.4.0
