+++ CHANGELOG.ru.md
@@ -3,8 +3,8 @@
 ## Unreleased
 
 ### Фичи
-- **Персональные API-токены (PAT):** залогиненный пользователь выпускает персональные API-токены (`POST`/`GET`/`DELETE /api/auth/tokens` + диалог *Персональные API-токены* в меню пользователя), передаваемые в заголовке `X-API-Key` — bearer-креды для не-браузерных клиентов (скрипты, будущий MCP-сервер) с сохранением роли (Casbin RBAC). Хранятся в виде хеша в БД снимков (таблица `api_tokens`, создаётся `dasha migrate`); least-privilege (запрашиваемая роль ≤ своей, по умолчанию `viewer`), anti-chaining (нельзя выпустить из-под PAT), одноразовый показ секрета, мгновенный отзыв, опциональный срок. Резолв токена кэшируется с коротким TTL, чтобы не ходить в БД на горячем пути авторизации; `last_used` обновляется не чаще раза за интервал на токен.
-- **MCP-коннектор (`dasha-mcp`):** отдельный read-only [MCP](https://modelcontextprotocol.io)-сервер поверх Dasha API для AI-ассистентов (Claude Desktop / Code, opencode, Cursor, …). 22 read-only tools (кластеры, ранжирование здоровья флота, instance info, health score/тренд/по-базам и рекомендации, top/активные/блокированные запросы, отчёт по запросам и сравнение снимков, индексы, крупнейшие таблицы, детальный разбор таблицы, репликация, анализ настроек, wait events, соединения, риск xid-wraparound, поиск по логам Yandex Cloud — с учётом rate limit: dedup по умолчанию, локальная валидация аргументов, чтобы невалидный запрос не сжигал слот лимита, а 429/501/504 переводятся в понятные модели подсказки) и 5 prompts (`diagnose_cluster`, `explain_health_score`, `find_index_opportunities`, `investigate_slow_queries`, `fleet_overview`). На каждом tool — MCP-аннотации read-only + closed-world, чтобы клиенты могли авто-аппрувить; сервер отдаёт инструкции по использованию для модели, а HTTP-режим шарит один кэш схем между per-token серверами. Работает по **stdio** (одна личность) или **HTTP/SSE** с per-user passthrough токена (общего серверного токена нет — Casbin RBAC сохраняется). Хардненинг: лимит размера ответа tool (слишком большой результат отклоняется, не режется в невалидный JSON), опциональные `limit`-аргументы у `describe_table` (партиции) и `connections` (сэмпл активности), хэшированный и ограниченный кэш серверов по токену, токены не логируются; rate-limit действует через вышестоящий per-identity лимитер Dasha. Двуязычная (en/ru, `--lang` / `DASHA_MCP_LANG`) встроенная база знаний в виде 3 MCP-ресурсов (`dasha://kb/health-rules` — каждое правило скоринга с порогами LOW/MED/HIGH и первыми действиями, синхронность с движком правил гарантирует sync-тест; `dasha://kb/wait-events` — глоссарий wait events; `dasha://kb/workflow` — сценарии «жалоба → цепочка» и правила бережности к API), а 5 prompts — линейные плейбуки (нумерованные шаги, один tool на шаг, критерий трактовки на каждом) — чтобы LLM без глубокой экспертизы PostgreSQL могли грамотно пользоваться инструментами, не замедляя сильные модели. Эксплуатационный паритет с бэкендом: каждый MCP-вызов логируется через zap в stderr (метод, tool, длительность — никогда аргументы или токены), HTTP-режим корректно дожидается in-flight запросов по SIGTERM/SIGINT, версия сервера прошивается из релизной сборки (та же схема internal/version + ldflags, что у бэкенда). Типизированный клиент к Dasha API из OpenAPI-спеки; multi-arch образ `deploy/images/Dockerfile.mcp`, публикуется в release-workflow; опциональный gated Helm Deployment + Service (`mcp.enabled`).
+- **Персональные API-токены (PAT):** залогиненный пользователь выпускает персональные API-токены (`POST`/`GET`/`DELETE /api/auth/tokens` + диалог *Персональные API-токены* в меню пользователя), передаваемые в заголовке `X-API-Key` — bearer-креды для не-браузерных клиентов (скрипты, будущий MCP-сервер) с сохранением роли (Casbin RBAC). Хранятся в виде хеша в БД снимков (таблица `api_tokens`, создаётся `dasha migrate`); least-privilege (запрашиваемая роль ≤ своей, по умолчанию `viewer`), anti-chaining (нельзя выпустить из-под PAT), одноразовый показ секрета, мгновенный отзыв, опциональный срок. Управление дополнительно закрыто ролёвкой `auth.pat_min_role` (`admin` по умолчанию, пока фича в обкатке; `viewer` открывает её всем залогиненным). Резолв токена кэшируется с коротким TTL, чтобы не ходить в БД на горячем пути авторизации; `last_used` обновляется не чаще раза за интервал на токен.
+- **MCP-коннектор (`dasha-mcp`):** отдельный read-only [MCP](https://modelcontextprotocol.io)-сервер поверх Dasha API для AI-ассистентов (Claude Desktop / Code, opencode, Cursor, …). 22 read-only tools (кластеры, ранжирование здоровья флота, instance info, health score/тренд/по-базам и рекомендации, top/активные/блокированные запросы, отчёт по запросам и сравнение снимков, индексы, крупнейшие таблицы, детальный разбор таблицы, репликация, анализ настроек, wait events, соединения, риск xid-wraparound, поиск по логам Yandex Cloud — с учётом rate limit: dedup по умолчанию, локальная валидация аргументов, чтобы невалидный запрос не сжигал слот лимита, а 429/501/504 переводятся в понятные модели подсказки) и 5 prompts (`diagnose_cluster`, `explain_health_score`, `find_index_opportunities`, `investigate_slow_queries`, `fleet_overview`). На каждом tool — MCP-аннотации read-only + closed-world, чтобы клиенты могли авто-аппрувить; сервер отдаёт инструкции по использованию для модели, а HTTP-режим шарит один кэш схем между per-token серверами. Работает по **stdio** (одна личность) или **HTTP/SSE** с per-user passthrough токена (общего серверного токена нет — Casbin RBAC сохраняется). Хардненинг: лимит размера ответа tool (слишком большой результат отклоняется, не режется в невалидный JSON), опциональные `limit`-аргументы у `describe_table` (партиции) и `connections` (сэмпл активности), хэшированный и ограниченный кэш серверов по токену, токены не логируются; rate-limit действует через вышестоящий per-identity лимитер Dasha. Двуязычная (en/ru, `--lang` / `DASHA_MCP_LANG`) встроенная база знаний в виде 3 MCP-ресурсов (`dasha://kb/health-rules` — каждое правило скоринга с порогами LOW/MED/HIGH и первыми действиями, синхронность с движком правил гарантирует sync-тест; `dasha://kb/wait-events` — глоссарий wait events; `dasha://kb/workflow` — сценарии «жалоба → цепочка» и правила бережности к API), а 5 prompts — линейные плейбуки (нумерованные шаги, один tool на шаг, критерий трактовки на каждом) — чтобы LLM без глубокой экспертизы PostgreSQL могли грамотно пользоваться инструментами, не замедляя сильные модели. Эксплуатационный паритет с бэкендом: каждый MCP-вызов логируется через zap в stderr (метод, tool, длительность — никогда аргументы или токены), HTTP-режим корректно дожидается in-flight запросов по SIGTERM/SIGINT, версия сервера прошивается из релизной сборки (та же схема internal/version + ldflags, что у бэкенда). Типизированный клиент к Dasha API из OpenAPI-спеки; multi-arch образ `deploy/images/Dockerfile.mcp`, публикуется в release-workflow; опциональный gated Helm Deployment + Service (`mcp.enabled`); чарт намеренно не предлагает общий fallback-токен — HTTP-режим строго per-user passthrough.
 
 ## v1.4.0
