+++ backend/internal/http/v1_common.go
@@ -6,6 +6,7 @@ import (
 	"fmt"
 
 	"github.com/dbulashev/dasha/gen/serverhttp"
+	"github.com/dbulashev/dasha/internal/auth"
 	"github.com/dbulashev/dasha/internal/config"
 	"github.com/dbulashev/dasha/internal/dto"
 	"github.com/dbulashev/dasha/internal/pkg/mapstruct"
@@ -24,10 +25,18 @@ func (s *Handlers) GetAuthInfo(
 
 	enableReset := s.cfg.EnableQueryStatsReset
 	// PATs belong to an individually-identifiable OIDC principal (shared static
-	// tokens cannot mint them) and require the api_tokens table to be migrated
-	// (else minting 500s and PAT auth fails closed). Only advertise the feature
-	// when both hold, so the frontend does not offer an unusable dialog.
-	patEnabled := s.cfg.Auth.Mode == config.AuthModeOIDC && s.storage.APITokensReady(ctx)
+	// tokens cannot mint them), are gated by auth.pat_min_role (admin-only by
+	// default while the feature matures) and require the api_tokens table to be
+	// migrated (else minting 500s and PAT auth fails closed). Only advertise the
+	// feature when all hold, so the frontend does not offer an unusable dialog.
+	callerRole := ""
+	if user := auth.UserFromContext(ctx); user != nil {
+		callerRole = user.Role
+	}
+
+	patEnabled := s.cfg.Auth.Mode == config.AuthModeOIDC &&
+		patMintAllowed(s.cfg.Auth.PATMinRole, callerRole) &&
+		s.storage.APITokensReady(ctx)
 	resp := serverhttp.GetAuthInfo200JSONResponse{
 		Mode:                  mode,
 		EnableQueryStatsReset: &enableReset,
